// shopping-cart (5)/shopping-cart/src/app/models/seller.model.ts
export class Seller
{
    // declare properties
    sellerId : number;
    sellerName : string;
    sellerMobile: string;
    sellerEmail: string;
    sellerCategory: string;
    sellerAddress: string;
    sellerPassword: string;
    sellerConfPass: string;
}

export class sellerVM{
    sellerId: number;
    sellerEmail: string;
    sellerPassword: string;
  }